requirements.txt
yt-dlp>=2023.0.0
openai>=1.0.0
stability-sdk>=0.8.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
requests>=2.28.0
pillow>=9.0.0
numpy>=1.22.0
tqdm>=4.64.0
